Seperate URI handler and sandbox
